Refactor TcpSocket and UdpSocket
