+++ src/openrct2/network/NetworkServerAdvertiser.cpp
@@ -24,7 +24,7 @@
 #    include "../world/Map.h"
 #    include "../world/Park.h"
 #    include "Http.h"
-#    include "UdpSocket.h"
+#    include "Socket.h"
 #    include "network.h"
 
 #    include <cstring>
